requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-redis==5.4.0
ollama==0.3.3
redis==5.0.8
django-cors-headers==4.4.0
python-decouple==3.8
whitenoise==6.7.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
pillow==10.4.0
celery==5.3.4
django-extensions==3.2.3
pytest==7.4.4
pytest-django==4.8.0
coverage==7.6.1
black==23.12.1
flake8==6.1.0
dj-database-url==2.2.0
django-heroku==0.3.1
djangorestframework-simplejwt==5.3.1